added reminder method

+++ src/main/java/OurDateTime.java
@@ -12,7 +12,7 @@ public class OurDateTime {
     private int minute;
     private String date;
     private String time;
-    private int CalculationFormat;
+    private long CalculationFormat;
 
 
     public OurDateTime (int year, int month, int day, int hour, int minute) {
@@ -102,7 +102,7 @@ public class OurDateTime {
         setCalculationFormat();
     }
 
-    public int getCalculationFormat() {
+    public long getCalculationFormat() {
         return CalculationFormat;
     }
